api/requirements.txt
Flask==3.1.0
flask_cors==5.0.1
flask_pymongo==3.0.1
flask_wtf==1.2.2
pytest==8.3.5
python-dotenv==1.0.1
# python_bcrypt==0.3.2
WTForms==3.2.1
bcrypt==4.3.0